Define a better wasm-js framebuffer interface
This lays the foundation for an emulated display device with config
registers to control how the display driver should interpret the
dimensions and bit depth of pixel data in the framebuffer.

Framebuffer capacity is 64KB, but the intent is that typically
video modes would use a config needing less than the whole buffer.