Rework demo html & css with bezel and drive bay
Previous layout was square cornered canvas directly on top of solid
color background. New layout has a dark round cornered bezel around
the canvas for the emulated screen. And, that is inset within a
larger rounded rectangle to make a console with an area for data
disc (cartridge?) slots to the right of the screen. Layout is
somewhat inspired by the Compaq Portable II 2650 luggable.