Speed up repaint
The repaint function was calling getImageData() to read old pixel
values from the canvas element, but then it was blindly overwriting
them. Also, that ImageData object could have been saved and reused,
but instead it was getting reallocated for each repaint.

This commit switches to a double buffer approach where the buffer
only gets reallocated when the screen configuration changes. On
localhost testing, these changes get the total load and wasm init
to repaint time down to about 89ms.